% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluscpv_fun.R
\name{pluscpv}
\alias{pluscpv}
\title{Use ASCAD or AMCP to estimate the chang-points.}
\usage{
pluscpv(Y, method = c("mcp", "scad"), p, lam.d)
}
\arguments{
\item{Y}{an autoregressive time series.}

\item{method}{the method to be used by mcp or scad.
See \code{\link[plus]{plus}} in R packages \pkg{plus} for details.}

\item{p}{an upper bound of autoregressive orders.}

\item{lam.d}{an parameter of the refining procedure. Suggest lam.d=0.05 if
the length of Y is smaller than 5000 and lam.d=0.01 otherwise.}
}
\value{
pluscpv returns an object of class "pluscpv".
An object of class "pluscpv" is a list containing the following components:

\item{variable.selection}{a matrix of variable selection.
Element of the	matrix is 1 if the variable is selected and 0	otherwise.}

\item{change.points}{estimators of change points.}
}
\description{
This function provides a novel and fast methodology for simultaneous multiple
structural break estimation and variable selection for nonstationary time series
models by using ASCAD or AMCP to estimate the chang-points by Jin, Shi and Wu (2011).
}
\references{
{Jin, B., Shi, X., and Wu, Y. (2013). A novel and fast methodology for
 simultaneous multiple structural break estimation and variable selection for
 nonstationary time series models. \emph{Statistics and Computing}, 23(2), 221-231.}
}
\seealso{
plus
}
